+++ tests/run_gui.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+"""
+Optima Core graphical user interface (GUI) script.
+Launches the GUI.
+"""
+
+from optimacore.gui import runGUI
+
+runGUI()